Fixed some of the failing unit tests. Some cleanup. Moved .editorconfig to parent folder.

+++ IoTEdgeModbus/Modules/AzureIoTEdgeModbus/Slave/Decoding/DiscreteBoolDecoder.cs
@@ -28,7 +28,7 @@
 
                     var addressIncrement = byteIndex * BitsInByte + bitIndex;
 
-                    var address = Convert.ToUInt16(operation.StartAddress) + addressIncrement;
+                    var address = Convert.ToInt32(operation.StartAddress) + addressIncrement;
 
                     result.Add(new DecodedValue(address, value.ToString()));
